Update protocol tests
