+++ tests/NATS.Client.Core2.Tests/NatsServerFixture.cs
@@ -5,6 +5,8 @@ namespace NATS.Client.Core2.Tests;
 // https://xunit.net/docs/shared-context#collection-fixture
 public class NatsServerFixture : IDisposable
 {
+    private int _next;
+
     public NatsServerFixture()
     {
         Server = NatsServerProcess.Start();
@@ -12,7 +14,11 @@ public class NatsServerFixture : IDisposable
 
     public NatsServerProcess Server { get; }
 
+    public int Port => new Uri(Server.Url).Port;
+
     public string Url => Server.Url;
+
+    public string GetNextId() => $"test{Interlocked.Increment(ref _next)}";
 
     public void Dispose()
     {
